commit: details below
Changed Page Order for page Navigation Menu/Bar

About Page: Revised writings more and added those writings to bulleted lists instead of paragraphs

Book Page: Added a border (same color blue as other borders) around the book prints paragraph and centered its text.

Styles: Added a new section for the About Page Border that I could not get to apply the border around. The lines in the code for that are 266 to 276.