commit: applied to all pages that were applicable
minor code clean up:

Made the table headers sentence case in the code, but kept the text uppercase via styling.

Made the bookmark text in the footer sentence case in the code but kept the text uppercase via styling.

Removed unnecessary line break from the vCard paragraph in the footer.

Added a line break to the paragraph for recommending A&P Master Imaging.